ControllerService/requirements.txt
flask
psycopg2-binary
pandas
flask_sqlalchemy
pika
google-api-python-client
